tests/requirements.txt
pyserial
ncurses
